Add setup.py for the new elasticdl_sdk package. Move DistributedOptimizer in this package at the first step.
