+++ scripts/build.sh
@@ -43,6 +43,11 @@ echo "Building the wheel for elasticdl_preprocessing."
 rm -rf ./build/lib
 python setup_preprocessing.py --quiet bdist_wheel --dist-dir ./build
 
+# Create elasticdl_sdk package
+echo "Building the wheel for elasticdl_sdk."
+rm -rf ./build/lib
+python setup_sdk.py --quiet bdist_wheel --dist-dir ./build
+
 # Create elasticdl_client package
 echo "Building the wheel for elasticdl_client."
 rm -rf ./build/lib
